add productsDevice.controller and remove other controllers
