--- src/controllers/Tablets.controller.ts
@@ -1,89 +1,0 @@
-'use strict';
-
-import { Request, Response } from 'express';
-import { ProductsService } from '../services/products.service';
-import { SortByOptions } from '../types/enums/Sorting';
-import { validateQueryParameters } from '../utils/helpers';
-import { ProductCategories } from '../types/enums/ProductCategories';
-import { ProductsDeviceService } from '../services/productsDevice.service';
-import { FindAllOptions } from '../types/findAllOptions';
-import { Op } from 'sequelize';
-
-const getAllTablets = async (req: Request, res: Response) => {
-  try {
-    const productsDeviceService = new ProductsDeviceService();
-    const productsService = new ProductsService();
-
-    const count = await productsDeviceService.count();
-
-    const {
-      limit = count,
-      offset = 0,
-      sortBy = SortByOptions.ID,
-      name,
-    } = req.query;
-
-    const { isSortByValid, isLimitValid, isOffsetValid } =
-      validateQueryParameters(+limit, +offset, sortBy as SortByOptions);
-
-    if (!isSortByValid || !isLimitValid || !isOffsetValid) {
-      res.sendStatus(400);
-
-      return;
-    }
-
-    const where: FindAllOptions['where'] = {};
-
-    where.category = ProductCategories.TABLETS;
-
-    if (name) {
-      where.name = { [Op.iLike]: `%${name}%` };
-    }
-
-    const tablets = await productsService.findAndCountAll({
-      limit: Number(limit),
-      offset: Number(offset),
-      sortBy: sortBy as SortByOptions,
-      where,
-    });
-
-    res.json(tablets);
-  } catch (error) {
-    console.error(error);
-
-    res.sendStatus(500);
-  }
-};
-
-const getOneTablet = async (req: Request, res: Response) => {
-  try {
-    const productsDeviceService = new ProductsDeviceService();
-
-    const { tabletId } = req.params;
-
-    if (Number(tabletId)) {
-      res.sendStatus(400);
-
-      return;
-    }
-
-    const foundTablet = await productsDeviceService.findById(tabletId);
-
-    if (!foundTablet) {
-      res.sendStatus(404);
-
-      return;
-    }
-
-    res.json(foundTablet);
-  } catch (error) {
-    console.error(error);
-
-    res.sendStatus(500);
-  }
-};
-
-export const tabletsController = {
-  getAllTablets,
-  getOneTablet,
-};
