+++ src/routes/phones.ts
@@ -1,11 +1,12 @@
 'use strict';
 
 import express from 'express';
-import { phonesController } from '../controllers/Phones.controller';
+import { devicesController } from '../controllers/ProductsDevice.controller';
+import { ProductCategories } from '../types/enums/ProductCategories';
 
 const router = express.Router();
 
-router.get('/', phonesController.getAllPhones);
-router.get('/:phoneId', phonesController.getOnePhone);
+router.get('/', devicesController.getAllDevices(ProductCategories.PHONES));
+router.get('/:deviceId', devicesController.getOneDevice);
 
 export default router;
